% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/restados.R
\name{restados}
\alias{restados}
\title{Resta dos valores numericos.}
\usage{
restados(x, y)
}
\arguments{
\item{x}{El primer operando}

\item{y}{El segundo operando}
}
\value{
La resta de los dos operandos, sin son números. \code{NaN}, en caso contrario
}
\description{
Esta funcion permite restar dos valores enteros o de coma flotante
}
\examples{
# La variable a la que se asigna el resultado de la funcion contendra
# el valor entero 6
seis <- restados(16,10)

# La funcion devolvera NaN, ya que los operandos informados no se
# corresponden con valores numericos
no.numero <- restados("No somos","numeros") # NaN
}
\keyword{condos}

