% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumados.R
\name{sumados}
\alias{sumados}
\title{Suma dos valores numericos.}
\usage{
sumados(x, y)
}
\arguments{
\item{x}{El primer operando}

\item{y}{El segundo operando}
}
\value{
La suma de los dos operandos, sin son números. \code{NaN} en caso
contrario
}
\description{
Esta función permite sumar dos valores enteros o de coma flotante
}
\examples{
# La variable que almacena el resultado retornado por la funcion contendra el
# valor entero 6
seis <- restados(4,2)

# La funcion, en este caso, devolvera el valor NaN, ya que los operandos
# informados no se corresponden con valores numericos
no.numero <- restados("No somos","numeros") # NaN
}
\keyword{condos}

